refactor(crypto): extract shared EncryptedFileStore and KeytarStore
Both totp-daemon/storage-provider.ts and secrets/keychain.ts had
nearly identical AES-256-GCM encrypted file storage and keytar
implementations. Extract shared classes into src/crypto/:

- EncryptedFileStore: AES-256-GCM with 12-byte IV, scrypt key
  derivation (32-byte output), per-file 16-byte salt, 0o700/0o600
  permissions
- KeytarStore: OS keychain access via lazy-loaded keytar module

Both consumers now delegate to the shared implementations, preserving
all security properties and external behavior.
